Sat, Jul 8, 2023, 3:48 PM +05:00

+++ half/halfs.c
@@ -4,25 +4,25 @@
 #include <cs50.h>
 #include <stdio.h>
 
-float half(float bill, float tax, int tip);
+float half(float bill, float tax, float tip);
 
 int main(void)
 {
     float bill_amount = get_float("Bill before tax and tip: ");
     float tax_percent = get_float("Sale Tax Percent: ");
-    int tip_percent = get_int("Tip percent: ");
+    float tip_percent = get_int("Tip percent: ");
 
     printf("You will owe $%.2f each!\n", half(bill_amount, tax_percent, tip_percent));
 }
 
 // TODO: Complete the function
-float half(float bill, float tax, int tip)
+float half(float bill, float tax, float tip)
 {
     bill = bill/2;
     tax = tax/2;
-    tip = (float) tip/2;
+    tip = tip/2;
     tax = (tax/100) * bill;
     tip = (tip/100) * bill;
-    printf("%f %f %i\n",bill,tax,tip);
+    printf("%f %f %f\n",bill,tax,tip);
     return 0.0;
 }
